Switch to Argon2i 1.3 instead of scrypt. This allows for more fine tuned control of the KDF which is useful on older computers. Also matches libsodium's default KDF. This includes all of Argon2, even though not all of it is used. A bit annoying, but simplifies tracking Argon2 changes.

+++ symops.h
@@ -21,7 +21,7 @@ struct hdr {
         unsigned char nonce[crypto_secretbox_NONCEBYTES];
         unsigned long long padded_len;
         long long rounds;
-        unsigned int mem;
+        long long mem;
         unsigned int p;
         int alg;
 };
